Fix handling of allowWarpEffects config and Warp Ward potion effect
allowWarpEffects is documented as "If false extra effects will only be
seen when using Pure Tear" however the current implementation makes it
so that the extra effects will **never** be seen. Using Pure Tear or any
of the other warp removal items this mod adds will queue effects, but
the queue will never be processed because the tick handler is never
registered. Fix that by always registering the tick handler, and
checking the config value for allowWarpEffects only when seeing if we
should do a periodic effect (as opposed to an effect caused by an item).

Additionally, the check for the Warp Ward potion effect (bypassable via
Pure Tear by default with disableRebound=false in config) had a
misplaced parenthesis that ended up causing effects to always activate
even when in Creative mode. This has an impact on testing worlds where
one may want to cheat in TC research by making the player then wait
through a gauntlet of crap for the queue to finish. Now, being in
Creative will properly suppress the effects although switching back to
Survival before the queue empties will still fire any remaining items in
there; I think that's probably fine but also fixable if desired.